Export methods to create DateError
Although the `DateError` struct was public, there was no way to create
one with the publicly available items. Now, the `date_error` and
`date_result` items are public, so users can create `DateError`s.

This can be useful, for example, when the user wants a function to
return a `DateResult` but performs some additional validation in
addition to the validation performed by the `parse_duration`
function. Now, users can perform the validation and return a
`DateError` if the validation fails.